audio_recorder/requirements.txt
pyaudio==0.2.13
openai-whisper==20240930 